+++ api/src/resolvers/campaigns.js
@@ -1,3 +1,0 @@
-//import { permitAdmin } from './permissions'
-
-export const campaignsMutations = {}