+++ web-react/src/pages/directory/display/ReadQueries.js
@@ -308,6 +308,10 @@ export const DISPLAY_BACENTA = gql`
         id
         name
         stream_name
+        council {
+          id
+          name
+        }
       }
       leader {
         id
@@ -430,6 +434,11 @@ export const DISPLAY_COUNCIL = gql`
       activeFellowshipCount
       memberCount
       pastorCount
+
+      stream {
+        id
+        name
+      }
       constituencies(options: { limit: 5 }) {
         id
         name
@@ -446,14 +455,6 @@ export const DISPLAY_COUNCIL = gql`
         lastName
         fellowship {
           id
-          bacenta {
-            id
-
-            constituency {
-              id
-              name
-            }
-          }
         }
       }
       leader {
@@ -486,6 +487,7 @@ export const DISPLAY_STREAM = gql`
     streams(where: { id: $id }, options: { limit: 1 }) {
       id
       name
+      target
       constituencyCount
       activeBacentaCount
       activeFellowshipCount