+++ web-react/src/pages/services/banking/self-banking/FellowshipSelfBanking.jsx
@@ -1,22 +1,23 @@
-import { useMutation, useQuery } from '@apollo/client'
+import { useQuery } from '@apollo/client'
 import { HeadingPrimary } from 'components/HeadingPrimary/HeadingPrimary'
 import PlaceholderCustom from 'components/Placeholder'
 import { ChurchContext } from 'contexts/ChurchContext'
-import { alertMsg, throwErrorMsg } from 'global-utils'
+import { throwErrorMsg } from 'global-utils'
 import { parseDate } from 'date-utils'
 import React, { useContext, useState } from 'react'
-import { Button, Card, Col, Container, Row, Spinner } from 'react-bootstrap'
+import { Card, Col, Container, Row } from 'react-bootstrap'
 import { useNavigate } from 'react-router'
 import { FELLOWSHIP_BANKING_SLIP_QUERIES } from '../../ServicesQueries'
 import HeadingSecondary from 'components/HeadingSecondary'
 import { capitalise } from 'global-utils'
-import { CONFIRM_OFFERING_PAYMENT } from './bankingQueries'
 import usePopup from 'hooks/usePopup'
 import Popup from 'components/Popup/Popup'
+import ConfirmPaymentButton from './components/button/ConfirmPayment'
 
 const FellowshipSelfBanking = () => {
   const { fellowshipId, clickCard } = useContext(ChurchContext)
   const { isOpen, togglePopup } = usePopup()
+  const [confirmService, setConfirmService] = useState(null)
   const navigate = useNavigate()
   const { data, loading, error, refetch } = useQuery(
     FELLOWSHIP_BANKING_SLIP_QUERIES,
@@ -24,8 +25,6 @@ const FellowshipSelfBanking = () => {
       variables: { fellowshipId: fellowshipId },
     }
   )
-  const [sending, setSending] = useState(false)
-  const [ConfirmOfferingPayment] = useMutation(CONFIRM_OFFERING_PAYMENT, {})
   const fellowship = data?.fellowships[0]
   const placeholder = ['', '', '']
   throwErrorMsg(error)
@@ -42,6 +41,19 @@ const FellowshipSelfBanking = () => {
       <HeadingSecondary loading={loading}>
         Please click to bank any of these services
       </HeadingSecondary>
+
+      {isOpen && (
+        <Popup handleClose={togglePopup}>
+          <div>
+            Your transaction status is pending please press this button to
+            confirm the status
+          </div>
+          <div className="d-grid gap-2">
+            <ConfirmPaymentButton service={confirmService} refetch={refetch} />
+          </div>
+        </Popup>
+      )}
+
       {data?.fellowships[0].services.map((service, index) => {
         if (service.noServiceReason || service.bankingSlip) {
           return null
@@ -49,52 +61,18 @@ const FellowshipSelfBanking = () => {
 
         return (
           <>
-            {isOpen && (
-              <Popup>
-                <div>
-                  Your transaction status is pending please press this button to
-                  confirm the status
-                </div>
-                <div className="d-grid gap-2">
-                  <Button
-                    variant="secondary"
-                    size="lg"
-                    className="p-3 mt-5"
-                    onClick={() => {
-                      setSending(true)
-                      ConfirmOfferingPayment({
-                        variables: {
-                          serviceRecordId: service?.id,
-                          stream_name: service?.stream_name,
-                        },
-                      })
-                        .then(() => {
-                          alertMsg('Payment Confirmed Successfully')
-                          navigate('/self-banking/receipt')
-                        })
-                        .catch((error) => {
-                          togglePopup()
-                          throwErrorMsg(error)
-                        })
-                        .then(() => {
-                          refetch({ fellowshipId: fellowshipId })
-                          setSending(false)
-                        })
-                    }}
-                  >
-                    Confirm Transaction{' '}
-                    {sending && <Spinner animation="grow" size="sm" />}
-                  </Button>
-                </div>
-              </Popup>
-            )}
             <Card
               key={index}
               className="mb-2"
               onClick={() => {
+                setConfirmService(service)
                 clickCard(service)
                 if (service?.transactionStatus === 'pending') {
                   togglePopup()
+                  return
+                }
+
+                if (service?.transactionStatus === 'success') {
                   return
                 }
                 navigate('/services/fellowship/self-banking/pay')